restructurer-ma-doc
