+++ service-products/api/docs/categoryswagger.js
@@ -1,0 +1,111 @@
+/**
+ * @swagger
+ * components:
+ *   schemas:
+ *     Category:
+ *       type: object
+ *       required:
+ *         - name
+ *       properties:
+ *         id:
+ *           type: integer
+ *           description: L'ID unique de la catégorie
+ *         name:
+ *           type: string
+ *           description: Nom de la catégorie
+ *       example:
+ *         id: 1
+ *         name: "Vêtements"
+ */
+
+/**
+ * @swagger
+ * /api/categories:
+ *   get:
+ *     summary: Récupère toutes les catégories
+ *     tags: [Categories]
+ *     responses:
+ *       200:
+ *         description: Liste de toutes les catégories
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: array
+ *               items:
+ *                 $ref: '#/components/schemas/Category'
+ *       500:
+ *         description: Erreur serveur
+ */
+
+/**
+ * @swagger
+ * /api/categories:
+ *   post:
+ *     summary: Crée une nouvelle catégorie
+ *     tags: [Categories]
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             $ref: '#/components/schemas/Category'
+ *     responses:
+ *       201:
+ *         description: Catégorie créée avec succès
+ *         content:
+ *           application/json:
+ *             schema:
+ *               $ref: '#/components/schemas/Category'
+ *       500:
+ *         description: Erreur serveur
+ */
+
+/**
+ * @swagger
+ * /api/categories/{id}:
+ *   put:
+ *     summary: Met à jour une catégorie existante
+ *     tags: [Categories]
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         schema:
+ *           type: integer
+ *         required: true
+ *         description: ID de la catégorie à mettre à jour
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             $ref: '#/components/schemas/Category'
+ *     responses:
+ *       200:
+ *         description: Catégorie mise à jour avec succès
+ *       404:
+ *         description: Catégorie non trouvée
+ *       500:
+ *         description: Erreur serveur
+ */
+
+/**
+ * @swagger
+ * /api/categories/{id}:
+ *   delete:
+ *     summary: Supprime une catégorie
+ *     tags: [Categories]
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         schema:
+ *           type: integer
+ *         required: true
+ *         description: ID de la catégorie à supprimer
+ *     responses:
+ *       200:
+ *         description: Catégorie supprimée avec succès
+ *       404:
+ *         description: Catégorie non trouvée
+ *       500:
+ *         description: Erreur serveur
+ */
